+++ .eslintrc.js
@@ -5,6 +5,10 @@ module.exports = {
     'standard',
     'promise'
   ],
+  rules: {
+    'no-control-regex': 0,
+    'no-global-assign': 0
+  },
   env: {
     node: true,
     jest: true